requirements.txt
pycoingecko~=1.4.0
urllib3>=1.26.2
discord.py>=1.6.0